Update README.md
The main idea is to create an AI-driven app that helps users identify edible plants and seasonal ingredients in nature. By taking pictures of the plants they find, users can quickly learn whether the plant is edible or toxic. The app also informs users about which ingredients are in season in their location and how abundant they are in the wild. Combining plant identification and seasonal information, users can create sustainable, local, and seasonal meals.